Control.Monad.Algebraic

MonadIO

class Monad m => MonadIO m where

Monads in which IO computations may be embedded. Any monad built by applying a sequence of monad transformers to the IO monad will be an instance of this class.

Instances should satisfy the following laws, which state that liftIO is a transformer of monads:

liftIO . return = return

liftIO (m >>= f) = liftIO m >>= (liftIO . f)

Methods

liftIO :: IO a -> m a

Lift a computation from the IO monad.

MonadUIO

class MonadUIO m where

The class of Monads that perform I/O without throwing synchronous exceptions.

Methods

liftUIO :: UIO a -> m a

liftIOSafely :: (MonadThrow m, MonadUIO m) => IO a -> m a

Lift an IO computation into m. If the given IO computation throws synchronous exceptions, these will be explicitly thrown under the MonadThrow effect.

MonadThrow

class Monad m => MonadThrow m where

A class for monads in which exceptions may be thrown.

Instances should obey the following law:

throwM e >> x = throwM e

In other words, throwing an exception short-circuits the rest of the monadic computation.

Methods

throwM :: Exception e => e -> m a

Throw an exception. Note that this throws when this action is run in the monad m, not when it is applied. It is a generalization of Control.Exception's throwIO.

Should satisfy the law:

throwM e >> f = throwM e

catch :: (Exception e, MonadThrow m) => CatchT m a -> (e -> m a) -> m a

try :: CatchT m a -> m (Either SomeException a)

MonadError

class Monad m => MonadError e m | m -> e where

Methods

throwError :: e -> m a

catchError :: Monad m => ExceptT e m a -> (e -> m a) -> m a

MonadReader

class Monad m => MonadReader r m | m -> r where

Minimal complete definition

Nothing

Methods

ask :: m r

reader :: (r -> a) -> m a

local :: MonadReader r m => (r -> r) -> ReaderT r m a -> m a

MonadState

class Monad m => MonadState s m | m -> s where

Minimal definition is either both of get and put or just state

Minimal complete definition

state | get, put

Methods

get :: m s

Return the state from the internals of the monad.

put :: s -> m ()

Replace the state inside the monad.

state :: (s -> (a, s)) -> m a

Embed a simple state action into the monad.

MonadWriter

class (Monoid w, Monad m) => MonadWriter w m | m -> w where

Minimal complete definition

Nothing

Methods

writer :: (a, w) -> m a

writer (a,w) embeds a simple writer action.

tell :: w -> m ()

tell w is an action that produces the output w.

listen :: MonadWriter w m => WriterT w m a -> m (a, w)